+++ bin/sam-build.mjs
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import { cli_run, sam_build_cli, } from '../lib/index.js'
+import { cli_run, sam_build_cli, } from '../src/index.js'
 cli_run(async () => {
 	const arg_a = process.argv.slice(2)
 	return sam_build_cli(arg_a, {