+++ bin/sam-local-start-api.mjs
@@ -1,9 +1,10 @@
 #!/usr/bin/env node
-import { cli_run, sam_local_start_api_cli, } from '../lib/index.js'
+import { import_meta_env_ } from '@ctx-core/env'
+import { cli_run, sam_local_start_api_cli, } from '../src/index.js'
 cli_run(async () => {
 	const arg_a = process.argv.slice(2)
 	return sam_local_start_api_cli(arg_a, {
 		'--template': 'stack-template.yaml',
-		'--port': process.env.PORT,
+		'--port': import_meta_env_().PORT,
 	})
-})
+}).then()