+++ src/stage_/index.js
@@ -1,8 +1,9 @@
+import { import_meta_env_ } from '@ctx-core/env'
 import { flag_value_ } from '../flag_value_/index.js'
 export function stage_(arg_a) {
 	const stage =
 		flag_value_(arg_a, '--stage')
-		|| process.env.NODE_ENV
+		|| import_meta_env_().NODE_ENV
 		|| 'development'
 	return stage
 }